refactor(sync): move the chunk round-trip out of the orchestrator
The third cut of #1777, and the core of a run: `ChunkDispatcher`
(`services/library/chunk_dispatcher.py`) now drives one work unit's
apply as a sequence of durable chunk round-trips — emit a chunk to the
frontend, wait for its acknowledgement against the heartbeat clock,
commit it through the reporter, and decide what becomes of a chunk whose
acknowledgement never comes.

Six methods moved verbatim: `_apply_unit_in_chunks` (public as
`apply_unit_in_chunks`, a peer calls it), the wait and the give-up half
(`_wait_for_unit_complete` / `_abandon_active_chunk`), and the two
final-chunk stamp builders, which are per-chunk facts evaluated inside
the loop against `chunk_index == chunk_count - 1` rather than the run's
finalisation. The three constants only that code used moved with it.

The absences in `ChunkDispatcherConfig` are the contract: no
Unit-of-Work factory, no event loop, no settings, no plugin directory,
and none of the confined seams. The dispatcher opens no transaction and
performs no I/O of its own — every durable write goes through the
reporter, every message through the injected emitter. The run's
lifecycle stays with the orchestrator too: the dispatcher may request a
cancel, but `try_begin_run` and `finish_run` are not its to call.

`_clip_cover_refreshes` did not come along. It touches no cover and
downloads nothing; every symbol it uses is renderer-budget arithmetic
over a reading the monitor produced a few lines earlier, so it is
`SessionBudgetMonitor.clip_cover_refreshes` now and the loop asks the
monitor two plain questions instead of doing the sums itself.

Two things are stated where they were only implicit. The comment on the
per-chunk identity stamping now says that those assignments must precede
the emit and what happens if they do not — a fast frontend's ack is
rejected as stray, the wait hangs, and the run dies of heartbeat timeout
sixty seconds later with the chunk stashed. And the platform stamp's two
ends now sit in two modules: the DELETE stayed at its call site, where
the argument for its ordering lives, while the re-stamp rides the final
chunk's commit UoW; each end names the other, and the register says so.

The orchestrator's module docstring claimed the heartbeat clock was
"refreshed on every progress emission", which was never true —
`emit_progress` is not among that field's writers. Corrected while
passing. `sleeper` left the orchestrator's config with the wait it
served; nothing else there used it.

Counted size 1192 → 954, back under the threshold, so the allowlist
entry is deleted rather than lowered.

Step of #1777.
