+++ py_modules/services/library/sync_orchestrator.py
@@ -2,15 +2,18 @@
 
 Owns every async path the user triggers from the QAM that mutates in-flight
 sync state: starting and cancelling syncs, computing a preview (read-only),
-and dispatching the per-unit sync pipeline on apply. The heartbeat clock —
-refreshed on every progress emission and inspected by per-unit waits — lives
-here too. Progress emission also lives here — sub-services that need to
+and dispatching the per-unit sync pipeline on apply. The heartbeat clock is
+stamped here — when a run is admitted, and by the frontend's ``sync_heartbeat``
+route — while the wait that inspects it moved out with the chunk round-trip it
+guards. Progress emission also lives here — sub-services that need to
 surface progress receive the orchestrator's ``emit_progress`` callback
 through their config. Anything that fetches ROMs belongs in
 :class:`LibraryFetcher`; anything that finalises shortcuts after the apply
 completes belongs in :class:`SyncReporter`; Steam's renderer memory belongs
 in :class:`SessionBudgetMonitor`; a single ROM's launch facts — where its
-file is and what runs it — belong in :class:`ShortcutLaunchResolver`; reading
+file is and what runs it — belong in :class:`ShortcutLaunchResolver`; driving
+one unit's built delta to the frontend and back — emit a chunk, wait for its
+ack, commit it — belongs in :class:`ChunkDispatcher`; reading
 the registry and the completion stamps into the projections these decisions
 are made against belongs in :class:`LocalLibraryReader`. What stayed here of that
 last group is the platform stamp's DELETE — a write, and a step of the apply
@@ -27,22 +30,12 @@ import asyncio
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any
 
-from domain.collection_sync_state import CollectionSyncState
 from domain.cover_refresh import count_cover_refreshes
-from domain.platform_sync_state import PlatformSyncState
 from domain.preview_delta import PreviewDelta
-from domain.session_budget import (
-    CLIFF_KB,
-    COVER_TRANSIENT_KB,
-    EFFECTIVE_CEILING_KB,
-    chunk_worst_cost_kb,
-    post_run_advisory,
-    session_memory_delta,
-)
+from domain.session_budget import post_run_advisory, session_memory_delta
 from domain.shortcut_data import build_shortcuts_data
 from domain.sibling_group import compute_component_group_keys
 from domain.sibling_resolution import AUTO_REGION
-from domain.sync_chunking import build_unit_chunks, wire_shortcuts
 from domain.sync_diff import (
     BIND_ROM_ID_KEY,
     classify_roms,
@@ -64,6 +57,7 @@ if TYPE_CHECKING:
     from domain.work_unit import WorkUnit
     from lib.late_binding import LateBinding
     from services.library._state import LibrarySyncStateBox
+    from services.library.chunk_dispatcher import ChunkDispatcher
     from services.library.fetcher import LibraryFetcher
     from services.library.local_library_reader import LocalLibraryReader
     from services.library.reporter import SyncReporter
@@ -72,7 +66,6 @@ if TYPE_CHECKING:
         ArtworkManager,
         Clock,
         EventEmitter,
-        Sleeper,
         UnitOfWorkFactory,
         UuidGen,
     )
@@ -85,23 +78,6 @@ _SYNC_CANCELLED = "Sync cancelled"
 # report "(interrupted)" instead of "(cancelled)" for a crash.
 _SYNC_INTERRUPTED = "Sync interrupted (Steam UI stopped responding)"
 _PREVIEW_MAX_AGE_SECONDS = 1800  # 30 minutes — preview snapshots stale beyond this
-
-# Per-unit heartbeat-based timeout. If the frontend stops calling
-# ``sync_heartbeat`` for this many seconds while the orchestrator is
-# waiting for ``report_unit_results``, the wait is treated as a
-# recoverable cancellation — the in-flight unit is dropped and the
-# next sync resumes via the incremental-skip path.
-_UNIT_HEARTBEAT_TIMEOUT_SEC = 60.0
-# Polling cadence the wait loop uses while watching the heartbeat
-# clock. Kept short so cancel propagation feels responsive without
-# burning CPU.
-_UNIT_WAIT_POLL_SEC = 1.0
-# Emitted-shortcut count per apply chunk. A unit's emitted shortcuts are split
-# into chunks of about this many entries, each emitted → acked → committed
-# durably before the next, so a mid-unit CEF crash forfeits only the in-flight
-# chunk. A chunk may overflow this to keep a sibling group whole (see
-# :func:`domain.sync_chunking.build_unit_chunks`).
-_APPLY_CHUNK_SIZE = 200
 
 
 def _collection_membership_key(unit: WorkUnit) -> tuple[str, str]:
@@ -121,7 +97,7 @@ class SyncOrchestratorConfig:
     """Frozen wiring bundle handed to ``SyncOrchestrator.__init__``.
 
     Holds runtime infrastructure (loop, logger), event emitter, the
-    Clock/UuidGen/Sleeper test seams, the SQLite Unit-of-Work factory
+    Clock/UuidGen test seams, the SQLite Unit-of-Work factory
     (the transactional seam over the ``roms`` / ``sync_runs`` repositories
     the lifecycle writes through), the plugin-dir reference for shortcut
     data construction, the shared
@@ -140,10 +116,15 @@ class SyncOrchestratorConfig:
     classify baseline, the per-unit bound-row projection, the unstamped-platform
     count, the resident sibling-group keys, and the stale-row scan — each in its
     own short read Unit of Work, where the fetcher asks the same kinds of
-    question of RomM. The ``session_budget`` seam owns every
-    renderer-heap reading and verdict: at each chunk boundary the orchestrator asks
-    it whether applying the chunk would exhaust Steam's per-session heap budget, and
-    the run pauses itself when it would (#1383).
+    question of RomM. The ``chunk_dispatcher`` peer takes the unit's built delta
+    the rest of the way: it emits each chunk to the frontend, waits out the
+    heartbeat clock for the ack, and commits the chunk through the reporter. The
+    ``session_budget`` seam owns every
+    renderer-heap reading and verdict, which the dispatcher asks at each chunk
+    boundary — whether applying the chunk would exhaust Steam's per-session heap
+    budget, and how much headroom is left for the chunk's additive cover work
+    (#1383); the terminal memory delta this module reports is drawn from the same
+    seam.
     """
 
     settings: dict[str, Any]
@@ -153,7 +134,6 @@ class SyncOrchestratorConfig:
     emit: EventEmitter
     clock: Clock
     uuid_gen: UuidGen
-    sleeper: Sleeper
     uow_factory: UnitOfWorkFactory
     sync_state_box: LibrarySyncStateBox
     fetcher: LibraryFetcher
@@ -162,6 +142,7 @@ class SyncOrchestratorConfig:
     shortcut_launch_resolver: ShortcutLaunchResolver
     local_library_reader: LocalLibraryReader
     session_budget: SessionBudgetMonitor
+    chunk_dispatcher: ChunkDispatcher
 
 
 @dataclass(frozen=True)
@@ -193,7 +174,6 @@ class SyncOrchestrator:
         self._emit = config.emit
         self._clock = config.clock
         self._uuid_gen = config.uuid_gen
-        self._sleeper = config.sleeper
         self._uow_factory = config.uow_factory
         self._sync_state = config.sync_state_box
         self._fetcher = config.fetcher
@@ -202,6 +182,7 @@ class SyncOrchestrator:
         self._shortcut_launch_resolver = config.shortcut_launch_resolver
         self._local_library_reader = config.local_library_reader
         self._session_budget = config.session_budget
+        self._chunk_dispatcher = config.chunk_dispatcher
 
     # ── Sync control ─────────────────────────────────────────────
 
@@ -1082,7 +1063,9 @@ class SyncOrchestrator:
         # heartbeat-timeout before the final chunk must leave NO stamp, so the
         # skip gate can't honour a stale one over a half-applied platform (the
         # #1025 silent-gap regression). The final chunk re-writes the stamp on a
-        # clean finish. A fetch that failed raised before here (fetch failure ≠
+        # clean finish, and that half of the pair now lives one module over —
+        # ``ChunkDispatcher._build_final_platform_stamp``, which rides the final
+        # chunk's commit UoW. A fetch that failed raised before here (fetch failure ≠
         # apply started) and a cancel during fetch/artwork returned at a guard
         # above with the old stamp intact, so an unstarted apply keeps it. A
         # skipped platform returned before this point and keeps its stamp. Only
@@ -1115,7 +1098,7 @@ class SyncOrchestrator:
             membership = collection_memberships.get(_collection_membership_key(unit))
             collection_member_ids = membership.rom_ids if membership is not None else None
 
-        applied_count = await self._apply_unit_in_chunks(
+        applied_count = await self._chunk_dispatcher.apply_unit_in_chunks(
             unit,
             unit_index=unit_index,
             total_units=total_units,
@@ -1175,300 +1158,6 @@ class SyncOrchestrator:
             if int(rom["id"]) in cover_paths and (source := rom.get("path_cover_large") or rom.get("path_cover_small"))
         }
 
-    async def _apply_unit_in_chunks(
-        self,
-        unit: WorkUnit,
-        *,
-        unit_index: int,
-        total_units: int,
-        emitted: list[dict[str, Any]],
-        shortcuts_data: list[dict[str, Any]],
-        unit_roms: list[dict[str, Any]],
-        new_ids: set[int],
-        cover_refreshes: list[dict[str, int]] | None = None,
-        collection_member_ids: list[int] | None = None,
-    ) -> int:
-        """Emit → wait → commit the unit's DELTA shortcuts one durable chunk at a time.
-
-        ``emitted`` is the delta (new + changed + rebind) the frontend applies;
-        skipped-unchanged entries never reach here but their rows still ride the
-        chunks' ``rom_ids`` (routed to chunk 0's leftover by ``build_unit_chunks``).
-        The delta shortcuts are split into commit chunks processed one at a time
-        (emit → wait → commit durably → next), so a mid-unit CEF crash forfeits
-        only the in-flight chunk, not every prior chunk. ``chunk.rom_ids`` are the
-        chunk's fetched ROMs (its sibling groups' rows, plus chunk 0's skipped
-        leftover); a keyed lookup into the whole unit's live fetch yields the
-        chunk's commit subset. ``new_ids`` (the classified creates) prices each
-        chunk create-vs-update for the session-budget gate. ``cover_refreshes``
-        (the #1386 invalidation pass's ``{rom_id, app_id}`` list) rides the
-        unit's FIRST chunk payload, clipped to the budget headroom left after
-        that chunk's own projected cost — a big refresh list degrades to fewer
-        in-session tile refreshes (the grid files are already updated; a Steam
-        restart shows the rest), never to a run pause. Returns the running
-        count of shortcuts applied — a cancel or heartbeat timeout returns early
-        with the chunks committed so far.
-        """
-        box = self._sync_state
-        # One generation id per platform fetch. The run id serves: a platform is
-        # fetched at most once per run, so it identifies this platform's fetch
-        # uniquely, and every chunk of the unit shares it — unlike a clock reading,
-        # which differs per chunk and would leave the earlier chunks' rows stamped
-        # before the final chunk's completion stamp (#1504).
-        fetch_id = str(box.current_sync_id or "") if unit.type == "platform" else None
-        chunks = build_unit_chunks(emitted, shortcuts_data, _APPLY_CHUNK_SIZE)
-        roms_by_id = {r["id"]: r for r in unit_roms if "id" in r}
-        chunk_count = len(chunks)
-        applied_count = 0
-        for chunk_index, chunk in enumerate(chunks):
-            # A cancel landing in the inter-chunk window — after the prior chunk's
-            # commit but before this chunk's emit — discards the rest of the unit
-            # here, before any per-chunk mutation or emit. Without this the
-            # frontend would fully process another ~200-shortcut chunk (~2 min)
-            # whose ack the backend then rejects, orphaning those shortcuts until
-            # the next sync. Same cleanup as the mid-wait user-cancel branch.
-            if box.is_cancelling():
-                box.clear_active_unit()
-                return applied_count
-
-            # Session-budget gate (#1383): at every chunk boundary ask the monitor
-            # whether applying this chunk would cross Steam's per-session heap budget,
-            # and pause here — a clean chunk boundary — if it would. On pause the
-            # gate sets ``run_paused`` + ``interrupt_reason`` and requests
-            # cancel, so the check just below returns cleanly with the prior chunks
-            # committed — the terminal finalize then records the resumable ``paused``
-            # state (the deliberate sibling of a heartbeat timeout's
-            # ``interrupted``). Both modes are PREDICTIVE (RSS plus this chunk's
-            # worst-case cost) and differ only in the line the projection is
-            # measured against:
-            #  - Every LATER chunk projects against the effective ceiling
-            #    (``cliff - margin`` ≈ 2.2 GB), keeping the anti-thrash safety margin.
-            #  - The run's very FIRST chunk projects against the CLIFF itself
-            #    (``CLIFF_KB`` ≈ 2.45 GB). Forward progress must be guaranteed — the
-            #    run has to apply at least one chunk or it loops forever on a
-            #    no-progress pause — so the first chunk is allowed to spend into the
-            #    safety margin, but the predictive projection still stops it before
-            #    the crash line. Net effect: a resume proceeds only when this chunk's
-            #    worst-case peak stays below the cliff (≈ 1.95 GB for a full 200-item
-            #    chunk of cover-applying creates, each priced create + cover) and can
-            #    never be projected past it; at/above that it re-pauses with zero
-            #    progress and the banner directs the user to restart Steam.
-            # The chunk is priced by composition, so the gate needs its creates and
-            # updates apart. The frontend decides create-vs-update itself via its
-            # existing-shortcut scan; a small backend/frontend mismatch only ever
-            # overprices (worst-case safe).
-            creates = sum(1 for e in chunk.emitted if e["rom_id"] in new_ids)
-            updates = len(chunk.emitted) - creates
-            budget_limit_kb = CLIFF_KB if box.chunks_emitted_this_run == 0 else EFFECTIVE_CEILING_KB
-            rss_kb = await self._session_budget.maybe_pause_for_budget(
-                creates=creates, updates=updates, limit_kb=budget_limit_kb
-            )
-            if box.is_cancelling():
-                box.clear_active_unit()
-                return applied_count
-
-            # The #1386 cover-refresh list rides the unit's FIRST chunk, clipped to
-            # the budget headroom left after this chunk's own projected cost — the
-            # refreshes must never be the reason a run pauses, so they degrade to
-            # fewer in-session tile refreshes instead (grid files already updated).
-            chunk_cover_refreshes: list[dict[str, int]] = []
-            if chunk_index == 0 and cover_refreshes:
-                chunk_cover_refreshes = self._clip_cover_refreshes(
-                    cover_refreshes, rss_kb=rss_kb, creates=creates, updates=updates, limit_kb=budget_limit_kb
-                )
-
-            chunk_rows = [roms_by_id[rid] for rid in chunk.rom_ids if rid in roms_by_id]
-
-            # Fresh per-chunk coordination: a new event + identity (run + unit +
-            # chunk index) so the reporter validates each chunk's ack.
-            box.unit_complete_event = asyncio.Event()
-            box.last_unit_results = None
-            box.active_unit_id = unit.id
-            box.active_chunk_index = chunk_index
-            box.sync_last_heartbeat = self._clock.monotonic()
-            await self._emit(
-                "sync_apply_unit",
-                {
-                    "run_id": str(box.current_sync_id or ""),
-                    "unit_type": unit.type,
-                    "unit_id": unit.id,
-                    "unit_name": unit.name,
-                    "unit_index": unit_index,
-                    "total_units": total_units,
-                    "chunk_index": chunk_index,
-                    "chunk_count": chunk_count,
-                    "chunk_offset": chunk.offset,
-                    "unit_total": len(emitted),
-                    # Strip backend-internal keys (staged cover path, rebind target)
-                    # from the wire — the frontend fetches a created shortcut's cover
-                    # via get_artwork_base64(rom_id); the commit reads them from
-                    # pending_sync, which keeps the full entries.
-                    "shortcuts": wire_shortcuts(chunk.emitted),
-                    # Existing shortcuts whose server-side cover changed (#1386):
-                    # the frontend re-applies each via SetCustomArtworkForApp so
-                    # the tile refreshes in-session. Non-empty only on chunk 0.
-                    "cover_refreshes": chunk_cover_refreshes,
-                },
-            )
-            # Count this emit so the session-budget gate exempts only the very
-            # first chunk of the run (forward-progress guarantee, #1383).
-            box.chunks_emitted_this_run += 1
-
-            applied = await self._wait_for_unit_complete(unit, box.unit_complete_event)
-            if applied is None:
-                # The wait gave up — the reason (user cancel vs heartbeat timeout)
-                # decides whether this chunk's in-flight work is recoverable. Chunks
-                # committed before this one stay committed either way.
-                self._abandon_active_chunk(box, chunk_rows)
-                return applied_count
-
-            platform_stamp = self._build_final_platform_stamp(unit, chunk_index, chunk_count, fetch_id)
-            collection_stamp = self._build_final_collection_stamp(unit, chunk_index, chunk_count, collection_member_ids)
-
-            # Per-chunk commit: the reporter upserts every fetched ROM of this
-            # chunk into the ``roms`` aggregate (identity + version metadata,
-            # unbound for non-representatives) and binds only the acked
-            # representatives, stamping each ROM's cached ``rom_metadata`` in the
-            # same write UoW (Rom row first, metadata second — FK-safe).
-            # ``chunk_rows`` is this chunk's slice of the live RomM fetch — the
-            # source of ``metadatum`` — so each committed chunk is a crash-safe
-            # checkpoint. The final-chunk ``platform_stamp`` / ``collection_stamp``
-            # (whichever the unit type produces) rides the same UoW.
-            await self._reporter.get().commit_unit_results(
-                applied,
-                chunk_rows,
-                platform_stamp=platform_stamp,
-                collection_stamp=collection_stamp,
-                # The fetch generation for a PLATFORM unit's rows (#1504), passed on
-                # EVERY chunk so the whole unit shares the generation the final
-                # chunk's stamp records. A collection unit passes None — it spans
-                # platforms, and re-marking a foreign platform's row would drop it
-                # from that platform's counted rows.
-                fetch_id=fetch_id,
-            )
-            applied_count += len(applied)
-            # Only a COMMITTED chunk's items count as done (#1383): an emitted chunk
-            # whose ack never landed — a cancel or a heartbeat timeout — returns above,
-            # before this line, so the paused banner never over-reports.
-            box.run_done_items += len(applied)
-
-        box.clear_active_unit()
-        return applied_count
-
-    def _abandon_active_chunk(self, box: LibrarySyncStateBox, chunk_rows: list[dict[str, Any]]) -> None:
-        """Tear down or stash the in-flight chunk after its wait gave up.
-
-        A user cancel (box already CANCELLING) intentionally discards the chunk:
-        drop the whole-unit staging, null the event, and clear the unit + chunk
-        identity so a stray late ack can't commit it. A heartbeat timeout (box
-        still RUNNING) instead stashes THIS chunk (its run/unit/chunk identity +
-        rows) into ``abandoned_chunk`` via ``stash_abandoned_chunk`` — inert data
-        that survives the run's teardown — while leaving the whole-unit staging
-        live, so a late ``report_unit_results`` still commits the delivered
-        bindings instead of leaving orphan shortcuts (#1052 / #1367). It then
-        marks the run ``interrupted`` (the frontend went dark, not the user's
-        Cancel — so the terminal SyncRun write records ``interrupted``) and
-        requests the cancel that stops the chunk loop.
-        """
-        if box.is_cancelling():
-            box.clear_active_unit()
-        else:
-            box.stash_abandoned_chunk(chunk_rows)
-            box.run_interrupted = True
-            box.request_cancel()
-
-    def _clip_cover_refreshes(
-        self,
-        cover_refreshes: list[dict[str, int]],
-        *,
-        rss_kb: int | None,
-        creates: int,
-        updates: int,
-        limit_kb: int,
-    ) -> list[dict[str, int]]:
-        """Clip the #1386 cover-refresh list to the chunk's remaining budget headroom.
-
-        Each refresh is a ``SetCustomArtworkForApp`` push costing one transient
-        cover (:data:`domain.session_budget.COVER_TRANSIENT_KB`) of renderer heap
-        at the chunk's peak, on top of the chunk's own projected cost. The clip
-        keeps only as many refreshes as fit under ``limit_kb`` so the refreshes
-        can never push a chunk the gate just approved over the line; the
-        remainder is skipped gracefully — their grid files are already updated,
-        so a Steam restart shows them (the same degradation the budget gate uses
-        elsewhere). ``rss_kb`` ``None`` (measurement unavailable) fails open and
-        keeps the whole list, mirroring the gate itself.
-        """
-        if rss_kb is None:
-            return cover_refreshes
-        headroom_kb = limit_kb - rss_kb - chunk_worst_cost_kb(creates, updates)
-        allowance = max(0, headroom_kb // COVER_TRANSIENT_KB)
-        if allowance >= len(cover_refreshes):
-            return cover_refreshes
-        self._logger.info(
-            f"Session-budget headroom clips cover refreshes: applying {allowance} of "
-            f"{len(cover_refreshes)}; the remaining grid files are already updated and "
-            f"show after a Steam restart"
-        )
-        return cover_refreshes[:allowance]
-
-    def _build_final_platform_stamp(
-        self, unit: WorkUnit, chunk_index: int, chunk_count: int, fetch_id: str | None = None
-    ) -> PlatformSyncState | None:
-        """Build the completion stamp for a platform unit's FINAL chunk, else ``None``.
-
-        On the final chunk of a PLATFORM unit the stamp rides that chunk's commit
-        UoW so "platform fully synced" ⟺ "stamp exists" is atomic on a crash. The
-        stamp lets the next sync's incremental-skip gate skip this platform even
-        when the whole run is later cancelled (its library-wide ``last_sync`` never
-        advances). Only platform units carry a skip gate — collections have none,
-        so they are never stamped. A cancel or heartbeat timeout mid-unit returns
-        before the final chunk, so an incomplete platform is never stamped
-        (ADR-0023 / #1025).
-        """
-        if unit.type == "platform" and unit.slug and chunk_index == chunk_count - 1:
-            return PlatformSyncState.stamp(
-                platform_slug=unit.slug,
-                at=self._clock.now().isoformat(),
-                rom_count=unit.rom_count,
-                fetch_id=fetch_id,
-            )
-        return None
-
-    def _build_final_collection_stamp(
-        self,
-        unit: WorkUnit,
-        chunk_index: int,
-        chunk_count: int,
-        member_rom_ids: list[int] | None,
-    ) -> CollectionSyncState | None:
-        """Build the completion stamp for a standard/smart collection's FINAL chunk, else ``None``.
-
-        The collection sibling of :meth:`_build_final_platform_stamp` (#742). On
-        the final chunk of a standard/smart collection unit whose listing carried an
-        ``updated_at``, the stamp rides that chunk's commit UoW so "collection
-        fully synced" ⟺ "stamp exists" is atomic on a crash. ``member_rom_ids`` is
-        the collection's FULL membership (every member id, not just the applied
-        new_roms), which a future skip replays to rebuild the Steam-collection
-        map. Virtual collections carry no stamp (no stable
-        ``updated_at``), and a cancel or heartbeat timeout mid-unit returns before
-        the final chunk — an incomplete collection is never stamped (ADR-0023).
-        """
-        if (
-            unit.type == "collection"
-            and unit.collection_kind in ("standard", "smart")
-            and unit.collection_updated_at
-            and member_rom_ids is not None
-            and chunk_index == chunk_count - 1
-        ):
-            return CollectionSyncState.stamp(
-                collection_id=str(unit.id),
-                collection_kind=unit.collection_kind,
-                updated_at=unit.collection_updated_at,
-                completed_at=self._clock.now().isoformat(),
-                rom_count=unit.rom_count,
-                member_rom_ids=tuple(member_rom_ids),
-            )
-        return None
-
     async def _sync_platform_unit(
         self,
         unit: WorkUnit,
@@ -1523,36 +1212,6 @@ class SyncOrchestrator:
                 virtual_type=unit.virtual_type,
             )
         return unit_roms, skipped
-
-    async def _wait_for_unit_complete(self, unit: WorkUnit, event: asyncio.Event) -> dict[str, int] | None:
-        """Heartbeat-based wait for the active unit's frontend callback.
-
-        Returns the frontend-reported ``rom_id_to_app_id`` on success.
-        Returns ``None`` on timeout or cancel — the outer loop maps that
-        onto a recoverable cancellation. The wait poll polls the
-        heartbeat clock rather than ``asyncio.wait_for(timeout=...)``
-        because the frontend sends ``sync_heartbeat`` calls during long
-        per-unit applies (artwork download, Set* calls) and a 60s
-        absolute cap would still race those.
-        """
-        box = self._sync_state
-        while not event.is_set():
-            if box.is_cancelling():
-                self._logger.info(f"Per-unit cancel observed while waiting for unit {unit.name}")
-                return None
-            elapsed = self._clock.monotonic() - box.sync_last_heartbeat
-            if elapsed > _UNIT_HEARTBEAT_TIMEOUT_SEC:
-                self._logger.warning(f"Per-unit timeout: no heartbeat for {elapsed:.0f}s waiting on unit {unit.name}")
-                return None
-            try:
-                await self._sleeper.sleep(_UNIT_WAIT_POLL_SEC)
-            except asyncio.CancelledError:
-                self._logger.info(f"Per-unit wait cancelled for unit {unit.name}")
-                raise
-
-        results = box.last_unit_results or {}
-        box.last_unit_results = None
-        return results
 
     async def _finalize_per_unit(
         self,
